型注釈の追加，およびMap.vueスクリプト部分の別ファイルへの分離
